+++ src/encryption.rs
@@ -1,4 +1,4 @@
-// Copyright 2019 MaidSafe.net limited.
+// Copyright 2021 MaidSafe.net limited.
 //
 // This SAFE Network Software is licensed to you under The General Public License (GPL), version 3.
 // Unless required by applicable law or agreed to in writing, the SAFE Network Software distributed
@@ -6,22 +6,29 @@
 // KIND, either express or implied. Please review the Licences for the specific language governing
 // permissions and limitations relating to use of the SAFE Network Software.
 
-use crate::sequential::{Iv, Key};
-use crate::SelfEncryptionError;
+use crate::Error;
 use aes::Aes128;
 use block_modes::block_padding::Pkcs7;
 use block_modes::{BlockMode, Cbc};
+use bytes::Bytes;
 type Aes128Cbc = Cbc<Aes128, Pkcs7>;
 
-pub const KEY_SIZE: usize = 16;
-pub const IV_SIZE: usize = 16;
+pub(crate) const KEY_SIZE: usize = 16;
+pub(crate) const IV_SIZE: usize = 16;
 
-pub fn encrypt(data: &[u8], key: &Key, iv: &Iv) -> Result<Vec<u8>, SelfEncryptionError> {
+pub(crate) const HASH_SIZE: usize = 32;
+pub(crate) const PAD_SIZE: usize = (HASH_SIZE * 3) - KEY_SIZE - IV_SIZE;
+
+pub(crate) struct Pad(pub [u8; PAD_SIZE]);
+pub(crate) struct Key(pub [u8; KEY_SIZE]);
+pub(crate) struct Iv(pub [u8; IV_SIZE]);
+
+pub(crate) fn encrypt(data: Bytes, key: &Key, iv: &Iv) -> Result<Bytes, Error> {
     let cipher = Aes128Cbc::new_fix(key.0.as_ref().into(), iv.0.as_ref().into());
-    Ok(cipher.encrypt_vec(data))
+    Ok(Bytes::from(cipher.encrypt_vec(data.as_ref())))
 }
 
-pub fn decrypt(encrypted_data: &[u8], key: &Key, iv: &Iv) -> Result<Vec<u8>, SelfEncryptionError> {
+pub(crate) fn decrypt(encrypted_data: Bytes, key: &Key, iv: &Iv) -> Result<Bytes, Error> {
     let cipher = Aes128Cbc::new_fix(key.0.as_ref().into(), iv.0.as_ref().into());
-    Ok(cipher.decrypt_vec(encrypted_data)?)
+    Ok(Bytes::from(cipher.decrypt_vec(encrypted_data.as_ref())?))
 }
